+++ app/src/main/java/com/example/foosball/models/FoosmenBar.java
@@ -1,0 +1,5 @@
+package com.example.foosball.models;
+
+public class FoosmenBar {
+
+}