+++ app/src/main/java/com/example/foosball/models/Universe.java
@@ -1,0 +1,18 @@
+package com.example.foosball.models;
+
+import com.example.foosball.drawing.GameBoard;
+
+import android.os.Bundle;
+
+import android.os.Handler;
+
+import android.app.Activity;
+
+import android.view.View;
+
+import android.view.View.OnClickListener;
+
+import android.widget.Button;
+
+public class Universe {
+}